refactored matchFilter instead of adding new method 'multipleFilter'
